requirements.txt
coverage>=5.3,<6.0
dj-database-url>=0.5.0
Django>=3.1.0,<4.0.0
gunicorn>=20.0.0,<21.0.0
psycopg2-binary>=2.8.0,<3.0.0
selenium>=3.141.0,<4.0.0
whitenoise>=5.2.0,<6.0.0
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.9
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.3
selenium==3.141.0
sqlparse==0.4.2
urllib3==1.26.7
validate-email==1.3
whitenoise==5.3.0
djangorestframework==3.13.1
django-cors-headers==3.10.1